mscoree: WIP: load native CLR if available.
The basic idea here is to load the Microsoft CLR installed on the system if
it's available for those applications which have dependencies on
functionality that's only present in the CLR.

There are a couple of formerly stub entry points that also had to be added
to mscoree.dll to make the Microsoft CLR load properly, namely ordinal 24
(appears to always be loaded by ordinal -- I don't know its real name),
which vends an unknown COM interface used by the CLR, CloseCtrs, and
GetExecutableProcessHeap.

WIP because this commit currently only moves _CorExeMain and the exports
which were directly observed in the logs to call back into mscoree from the
CLR over to the indirect operations table. For completeness, the entire API
surface should be moved over to it.